+++ isitfit/cost/redshift/iterator.py
@@ -177,8 +177,12 @@ class BaseIterator:
 
     self.n_entry = len(list(self.iterate_core(True)))
 
-    msg_count = "Found a total of %i %s in %i region(s) (other regions do not hold any %s)"
-    logger.warning(msg_count%(self.n_entry, self.service_description, len(self.region_include), self.service_name))
+    if self.n_entry==0 and len(self.region_include)==0:
+      msg_count = "Found no %s"
+      logger.warning(msg_count%(self.service_description))
+    else:
+      msg_count = "Found a total of %i %s in %i region(s) (other regions do not hold any %s)"
+      logger.warning(msg_count%(self.n_entry, self.service_description, len(self.region_include), self.service_name))
 
     return self.n_entry